Create MSP430_Multiple_Channel_ADC
With this file, we can convert multiple analog channels in the MSP430 ADC10 module. 6 Potentiometers (or 6 potentiometers and the two axis of an analog joystick) are connected to the A0, A3, A4, A5, A6 & A7 inputs. The read values are output to the terminal at 19200 baud. 
DTC (Data Transfer Controller) property of the ADC10 module is used for multiple read. Please see the following video for detailed explanation:
https://youtu.be/Ev871bhGFt0